+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tech/machine/tile/TileBrickKiln.java
@@ -24,7 +24,7 @@ public class TileBrickKiln
   }
 
   @Override
-  protected double getFuelBurnTimeModifier(ItemStack fuel) {
+  public double getFuelBurnTimeModifier(ItemStack fuel) {
 
     return ModuleTechMachineConfig.BRICK_KILN.FUEL_BURN_TIME_MODIFIER;
   }