+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tech/machine/tile/TileStoneCrucible.java
@@ -44,7 +44,7 @@ public class TileStoneCrucible
   }
 
   @Override
-  protected double getFuelBurnTimeModifier(ItemStack fuel) {
+  public double getFuelBurnTimeModifier(ItemStack fuel) {
 
     return ModuleTechMachineConfig.STONE_CRUCIBLE.FUEL_BURN_TIME_MODIFIER;
   }